Changed product selection
This is a possible fix for issue where product variant at `configurableVariantIndex === 0` is displaying sku of a parent configurable product and not it's own.

Issue caused by the way Javascript handles Object and number 0 in a logical `&&`

This solution works both on configurable products with variants and on single product, but more explicit checks might be necessary.

Let me know if you want this fix improved